Git ignore list
The 'datasets' folder can be added to the '.gitignore' file along with
other temporary files resulting from Parboil runs. In this way
'datasets' won't be part of the repository, keeping Parboil size
limited. If not present, data sets are actually downloaded from a remote
AWS location when Parboil is run.